+++ src/js/controller/carManager/validation.js
